fix/refactor types pretty printer in code generator
