+++ pynestml/codegeneration/expressions_pretty_printer.py
@@ -21,7 +21,7 @@
 
 from typing import Tuple
 
-from pynestml.codegeneration.i_reference_converter import IReferenceConverter
+from pynestml.codegeneration.cpp_types_printer import CppTypesPrinter
 from pynestml.codegeneration.nestml_reference_converter import NestMLReferenceConverter
 from pynestml.meta_model.ast_expression import ASTExpression
 from pynestml.meta_model.ast_expression_node import ASTExpressionNode
@@ -40,8 +40,6 @@ class ExpressionsPrettyPrinter:
     """
 
     def __init__(self, reference_converter=None, types_printer=None):
-        # type: (IReferenceConverter,TypesPrinter) -> None
-        # todo by kp: this should expect a ITypesPrinter as the second arg
         if reference_converter is not None:
             self.reference_converter = reference_converter
         else:
@@ -50,7 +48,7 @@ class ExpressionsPrettyPrinter:
         if types_printer is not None:
             self.types_printer = types_printer
         else:
-            self.types_printer = TypesPrinter()
+            self.types_printer = CppTypesPrinter()
 
     def print_expression(self, node, prefix=''):
         """Print an expression.
@@ -157,22 +155,3 @@ class ExpressionsPrettyPrinter:
             ret.append(self.print_expression(arg, prefix=prefix))
 
         return tuple(ret)
-
-
-class TypesPrinter:
-    """
-    Returns a processable format of the handed over element.
-    """
-
-    @classmethod
-    def pretty_print(cls, element):
-        assert (element is not None), \
-            '(PyNestML.CodeGeneration.PrettyPrinter) No element provided (%s)!' % element
-        if isinstance(element, bool) and element:
-            return 'true'
-        elif isinstance(element, bool) and not element:
-            return 'false'
-        elif isinstance(element, int) or isinstance(element, float):
-            return str(element)
-        elif isinstance(element, str):
-            return element
