+++ pynestml/codegeneration/types_printer.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+#
+# types_printer.py
+#
+# This file is part of NEST.
+#
+# Copyright (C) 2004 The NEST Initiative
+#
+# NEST is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 2 of the License, or
+# (at your option) any later version.
+#
+# NEST is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with NEST.  If not, see <http://www.gnu.org/licenses/>.
+
+import abc
+
+
+class TypesPrinter(metaclass=abc.ABCMeta):
+    """
+    Returns a processable format of the handed over element.
+    """
+
+    @classmethod
+    @abc.abstractmethod
+    def pretty_print(cls, element):
+        pass
